# src/main/resources/public/swagger/swagger.yaml
openapi: '3.0.3'
info:
  title: baragym-backend
  version: 0.0.1
  description: Backend da aplicação BaraGym.
servers:
  - url: 'https://baragym-backend.herokuapp.com/api/v1'
  - url: 'http://localhost:8080/api/v1'
...
paths:
...
  /turmas:
    post:
      tags:
        - turma
      summary: Cria uma turma
      description: Cria uma turma a partir do nome e capacidade.
      security:
        - bearerAuthToken: [ ]
      requestBody:
        description: Corpo da requisição de criação de turma
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTurmaRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TurmaResponse'
        '400':
          description: Bad request. Falha na validação dos dados de entrada ou violação de regras de negócio.
        '401':
          description: Unauthorized. Retornado quando a requisição não envia as credenciais necessárias para identificação e autorização.
        '403':
          description: Forbidden. As credenciais fornecidas são insuficientes para realizar esta operação.
        '500':
          description: Erro interno
    get:
      tags:
        - turma
      summary: Obtém todas as turmas
      security:
        - bearerAuthToken: [ ]
      parameters:
        - name: page
          in: query
          description: |
            Número de página solicitado. <br/>
            **Atenção: os números de página inciam em `0`.**
          required: false
          schema:
            type: integer
            example: 0
            default: 0
        - name: size
          in: query
          description: Tamanho máximo da página de dados retornada.
          required: false
          schema:
            type: integer
            example: 20
            default: 10
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TurmaPageResponse'
        '400':
          description: Bad request. Falha na validação dos dados de entrada ou violação de regras de negócio.
        '401':
          description: Unauthorized. Retornado quando a requisição não envia as credenciais necessárias para identificação e autorização.
        '403':
          description: Forbidden. As credenciais fornecidas são insuficientes para realizar esta operação.
        '500':
          description: Erro interno
  /turmas/{turmaId}:
    get:
      tags:
        - turma
      summary: |
        Obtem uma turma
      description: |
        Apenas administradores (`ADMIN`).
      security:
        - bearerAuthToken: [ ]
      parameters:
        - name: turmaId
          in: path
          description: Identificador único da turma
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TurmaResponse'
        '400':
          description: Bad request. Falha na validação dos dados de entrada ou violação de regras de negócio.
        '401':
          description: Unauthorized. Retornado quando a requisição não envia as credenciais necessárias para identificação e autorização.
        '403':
          description: Forbidden. As credenciais fornecidas são insuficientes para realizar esta operação.
        '500':
          description: Erro interno
    put:
      tags:
        - turma
      summary: |
        Atualiza uma turma
      description: |
        Apenas administradores (`ADMIN`).
      security:
        - bearerAuthToken: [ ]
      parameters:
        - name: turmaId
          in: path
          description: Identificador único da turma
          required: true
          schema:
            type: string
      requestBody:
        description: Corpo da requisição de atualização da turma.
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTurmaRequest'
      responses:
        '200':
          description: OK
        '400':
          description: Bad request. Falha na validação dos dados de entrada ou violação de regras de negócio.
        '401':
          description: Unauthorized. Retornado quando a requisição não envia as credenciais necessárias para identificação e autorização.
        '403':
          description: Forbidden. As credenciais fornecidas são insuficientes para realizar esta operação.
        '500':
          description: Erro interno
    delete:
      tags:
        - turma
      summary: |
        Deleta uma turma
      description: |
        Apenas administradores (`ADMIN`).
      security:
        - bearerAuthToken: [ ]
      parameters:
        - name: turmaId
          in: path
          description: Identificador único da turma
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
        '400':
          description: Bad request. Falha na validação dos dados de entrada ou violação de regras de negócio.
        '401':
          description: Unauthorized. Retornado quando a requisição não envia as credenciais necessárias para identificação e autorização.
        '403':
          description: Forbidden. As credenciais fornecidas são insuficientes para realizar esta operação.
        '500':
          description: Erro interno
  
  ######################
  ##### ATIVIDADE ######
  ######################
...
components:
  securitySchemes:
    bearerAuthToken:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Token de acesso à aplicação.
  schemas:
...
    SortResponse:
      type: object
      properties:
        empty:
          type: boolean
          example: true
          description: Verifica se a lista está vazia ou não.
        sorted:
          type: boolean
          example: false
          description: Verifica se a lista está ordenada ou não.
        unsorted:
          type: boolean
          example: true
          description: Verifica se a lista está fora de ordenação ou não.
...
    CreateTurmaRequest:
      type: object
      properties:
        nome:
          type: string
          description: Nome da turma
          example: 'Sala 1'
        capacidade:
          type: number
          description: Capacidade da turma (pessoas)
          example: 10
    TurmaResponse:
      type: object
      properties:
        id:
          type: number
          description: identificador único da turma
          example: 12
        nome:
          type: string
          description: Nome da turma
          example: 'Sala 1'
        capacidade:
          type: number
          description: Capacidade da turma (pessoas)
          example: 10
    TurmaPageResponse:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/TurmaResponse'
        pageable:
          type: object
        last:
          type: boolean
          example: false
          description: Verifica se é a última página.
        totalElements:
          type: number
          example: 5
          description: Número de elementos totais.
        totalPages:
          type: number
          example: 5
          description: Número de páginas totais.
        size:
          type: number
          example: 1
          description: Número de elementos por página.
        number:
          type: number
          example: 0
          description: Número de da página.
        sort:
          $ref: '#/components/schemas/SortResponse'
        first:
          type: boolean
          example: true
          description: Verifica se é a primeira página.
        numberOfElements:
          type: number
          example: 5
          description: Número de elementos totais.
        empty:
          type: boolean
          example: false
          description: Verifica se a lista está vazia ou não.
...